generator: move uniform generators in uniform.go
